move experiments/datakit_testbed -> experiments/testbed
Drop the redundant datakit_ prefix. Rewrite import paths in moved
modules, tests/datakit_testbed/*, and scripts/datakit/run_source_sampling.py.
The tests/datakit_testbed/ directory keeps its name for now.
