+++ experiments/testbed/variants.py
@@ -32,10 +32,10 @@ from marin.processing.classification.deduplication.fuzzy_dups import FuzzyDupsAt
 from marin.processing.classification.deduplication.fuzzy_minhash import MinHashAttrData, compute_minhash_attrs
 from rigging.log_setup import configure_logging
 
-from experiments.datakit_testbed.mixture import tokenized_bucket_weights_step
-from experiments.datakit_testbed.sampler import build_testbed_steps
-from experiments.datakit_testbed.settings import TESTBED_TOKENIZER
-from experiments.datakit_testbed.train import run_testbed_config, testbed_tokenize
+from experiments.testbed.mixture import tokenized_bucket_weights_step
+from experiments.testbed.sampler import build_testbed_steps
+from experiments.testbed.settings import TESTBED_TOKENIZER
+from experiments.testbed.train import run_testbed_config, testbed_tokenize
 
 logger = logging.getLogger(__name__)
