+++ tests/datakit_testbed/test_sampler_pipeline.py
@@ -6,7 +6,7 @@
 import pytest
 from marin.datakit.sources import DatakitSource, all_sources
 
-from experiments.datakit_testbed.sampler import build_testbed_steps
+from experiments.testbed.sampler import build_testbed_steps
 
 _ALL = all_sources()
 _ALL_LIST = list(_ALL.values())
